requirements.txt
iocextract==1.13.1
parsel==1.5.2
urllib3==1.24.3
colorama>=0.4.3
flake8==3.7.9
tqdm>=4.46.1
pymongo==3.10.1
flask~=1.1.2
gitpython
python-magic==0.4.18
requests>=2.22.0
semver~=2.10.2
bitbucket-python
pip~=20.1.1
wheel~=0.34.2
nose~=1.3.7
setuptools~=40.0.0
pbr==5.4.5
bandit==1.6.2
njsscan==0.1.5
click~=7.1.2
python-levenshtein==0.12.0
packaging==20.4
lxml~=4.5.2
pandas~=1.2.5